program works now, but still doesn't converge to right value
I see no reason why this should be so, no computational reason I mean.
I am quite sure the program does what it is supposed to, and that the reason
for the errors lie in the fact that I explicitly demand in this program that
the off diagonal blocks are zero. In the other program I just assume that I
have no need for them and no knowledge, here I do have need for them and I set
them to zero, this is what I think causes the "to high" value of the energy.