I found out I've been doing it all wrong
I forgot to symmetrize the a=l terms correctly troughout the projection, so I have to start over.
The good thing is, I already did the W^l_{aa;aa}, and now the projection is only orthogonal when
the basis set is normalized, which is a good thing. So now I understand that what I had to change
to make the projection orthogonal, was not logical because the projection was not complete.