this is the first commit of the transfer branch
this is a branch made to transfer coupled objects to uncoupled objects to
be able to do check's between the different programs.

I added two classes, the uncoupled rxTPM and dDPM and have written and
tested the "uncouple" function. Eigenvalues are the same coupled and
uncoupled and with the right degenerecies, so this works.